@@ -1,11 +1,12 @@
-package ru.romanow.dictionary;
+package ru.romanow.dictionary
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
 
 @SpringBootApplication
-public class DictionaryApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(DictionaryApplication.class, args);
+object DictionaryApplication {
+    @JvmStatic
+    fun main(args: Array<String>) {
+        SpringApplication.run(DictionaryApplication::class.java, *args)
     }
 }
