@@ -1,12 +1,9 @@
-package ru.romanow.dictionary.mappers;
+package ru.romanow.dictionary.mappers
 
-import org.mapstruct.Mapper;
-import ru.romanow.dictionary.domain.LegoSetEntity;
-import ru.romanow.dictionary.mappers.config.MappingConfig;
-import ru.romanow.dictionary.models.LegoSet;
+import org.mapstruct.Mapper
+import ru.romanow.dictionary.domain.LegoSetEntity
+import ru.romanow.dictionary.mappers.config.MappingConfig
+import ru.romanow.dictionary.models.LegoSet
 
-@Mapper(config = MappingConfig.class)
-public interface LegoSetMapper
-        extends ModelMapper<String, LegoSetEntity, LegoSet> {
-
-}
+@Mapper(config = MappingConfig::class)
+interface LegoSetMapper : ModelMapper<String, LegoSetEntity, LegoSet>
