@@ -1,15 +1,8 @@
-package ru.romanow.dictionary.models;
+package ru.romanow.dictionary.models
 
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain = true)
-public class Series {
-
-    private String name;
-    private String type;
-    private String complexity;
-    private Integer age;
-
-}
+data class Series(
+    var name: String? = null,
+    var type: String? = null,
+    var complexity: String? = null,
+    var age: Int? = null
+)
