@@ -1,9 +1,6 @@
-package ru.romanow.dictionary.repository;
+package ru.romanow.dictionary.repository
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.romanow.dictionary.domain.SeriesEntity;
+import org.springframework.data.jpa.repository.JpaRepository
+import ru.romanow.dictionary.domain.SeriesEntity
 
-public interface LegoSetEntityRepository
-        extends JpaRepository<SeriesEntity, String> {
-
-}
+interface LegoSetEntityRepository : JpaRepository<SeriesEntity, String>
