@@ -1,9 +1,6 @@
-package ru.romanow.dictionary.repository;
+package ru.romanow.dictionary.repository
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.romanow.dictionary.domain.LegoSetEntity;
+import org.springframework.data.jpa.repository.JpaRepository
+import ru.romanow.dictionary.domain.LegoSetEntity
 
-public interface SeriesEntityRepository
-        extends JpaRepository<LegoSetEntity, String> {
-
-}
+interface SeriesEntityRepository : JpaRepository<LegoSetEntity, String>
