@@ -1,18 +1,17 @@
-package ru.romanow.dictionary;
+package ru.romanow.dictionary
 
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.ActiveProfiles;
-import ru.romanow.dictionary.config.DatabaseTestConfiguration;
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.context.annotation.Import
+import org.springframework.test.context.ActiveProfiles
+import ru.romanow.dictionary.config.DatabaseTestConfiguration
 
 @ActiveProfiles("test")
 @SpringBootTest
-@Import(DatabaseTestConfiguration.class)
-class DictionaryApplicationTest {
+@Import(DatabaseTestConfiguration::class)
+internal class DictionaryApplicationTest {
 
     @Test
-    void test() {
+    fun test() {
     }
-
 }
