@@ -1,10 +1,12 @@
-package ru.romanow.gateway.config.properties;
+package ru.romanow.gateway.config.properties
 
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConfigurationProperties
 
-@Data
 @ConfigurationProperties("application")
-public class ApplicationProperties {
-    private RoutesProperties routes;
+data class ApplicationProperties(
+    var routes: RoutesProperties
+)
+
+class RoutesProperties {
+    var dictionary: String? = null
 }
